refactor(step1): refactor project directory structure
