+++ src/main/java/qna/repository/entity/User.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.repository.entity;
 
 import qna.UnAuthorizedException;
