+++ src/main/java/qna/service/QnaService.java
@@ -7,6 +7,12 @@ import org.springframework.transaction.annotation.Transactional;
 import qna.CannotDeleteException;
 import qna.NotFoundException;
 import qna.domain.*;
+import qna.repository.AnswerRepository;
+import qna.repository.QuestionRepository;
+import qna.repository.entity.Answer;
+import qna.repository.entity.DeleteHistory;
+import qna.repository.entity.Question;
+import qna.repository.entity.User;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
